+++ src/rex.formbuilder/src/rex/formbuilder/__init__.py
@@ -3,6 +3,7 @@
 #
 
 
+from .action import *
 from .api import *
 from .widgets import *